Tests for single ops over websocket transport
