+++ server/integration-tests/src/test/java/io/smallrye/graphql/tests/client/dynamic/DynamicClientSingleOperationsTestBase.java
@@ -20,27 +20,19 @@ import java.util.concurrent.ExecutionException;
 import javax.json.JsonObject;
 
 import org.jboss.arquillian.container.test.api.Deployment;
-import org.jboss.arquillian.container.test.api.RunAsClient;
-import org.jboss.arquillian.junit.Arquillian;
 import org.jboss.arquillian.test.api.ArquillianResource;
 import org.jboss.shrinkwrap.api.ShrinkWrap;
 import org.jboss.shrinkwrap.api.asset.EmptyAsset;
 import org.jboss.shrinkwrap.api.spec.WebArchive;
-import org.junit.After;
-import org.junit.Before;
 import org.junit.Test;
-import org.junit.runner.RunWith;
 
 import io.smallrye.graphql.client.Response;
 import io.smallrye.graphql.client.core.Document;
 import io.smallrye.graphql.client.core.ScalarType;
 import io.smallrye.graphql.client.core.Variable;
 import io.smallrye.graphql.client.vertx.dynamic.VertxDynamicGraphQLClient;
-import io.smallrye.graphql.client.vertx.dynamic.VertxDynamicGraphQLClientBuilder;
 
-@RunWith(Arquillian.class)
-@RunAsClient
-public class DynamicClientTest {
+public abstract class DynamicClientSingleOperationsTestBase {
 
     @Deployment
     public static WebArchive deployment() {
@@ -52,19 +44,7 @@ public class DynamicClientTest {
     @ArquillianResource
     URL testingURL;
 
-    private static VertxDynamicGraphQLClient client;
-
-    @Before
-    public void prepare() {
-        client = (VertxDynamicGraphQLClient) new VertxDynamicGraphQLClientBuilder()
-                .url(testingURL.toString() + "graphql")
-                .build();
-    }
-
-    @After
-    public void cleanup() {
-        client.close();
-    }
+    VertxDynamicGraphQLClient client;
 
     @Test
     public void testOneQueryInOneOperationSync() throws ExecutionException, InterruptedException {
